refactor: move publishing result tests
